Fix some warnings from https://lgtm.com/projects/g/mfem/mfem

+++ general/array.hpp
@@ -1044,8 +1044,7 @@ void BlockArray<T>::Swap(BlockArray<T> &other)
 template<typename T>
 long BlockArray<T>::MemoryUsage() const
 {
-   return blocks.Size()*(mask+1)*sizeof(T) +
-          blocks.MemoryUsage();
+   return (unsigned int)blocks.Size()*(mask+1)*sizeof(T) + blocks.MemoryUsage();
 }
 
 template<typename T>
